Add support for LibreELEC settings to the settings overview.
Inspired by the Estuary commit [1] which only makes the LibreELEC window show when the LibreELEC service add-on is actually present.
See related Kodi forums thread [2] also.

Since Arctic Zephyr uses a two-column, five line layout for the settings menu, the LibreELEC settings entry would be hidden in a third column that
only shows when you click right a second time. To remedy this, this commit also rearranges a three-column, four line layout so the entries retain
the same size. 11 being an odd number means it looks, well, a bit odd, but at least it's showing and accessible.

The LibreELEC settings menu is not skinned since Arctic Zephyr does not support it by default, which limits its immediate usability. A PR fixing
that is more than welcome (not a priority).

[1] https://github.com/phil65/skin.estuary/commit/cbcee29ade2796c8c2a86e4afccfff3eb2a984e8
[2] https://forum.kodi.tv/showthread.php?tid=299073&pid=2470845#pid2470845

Signed-off-by: Stijn Segers <foss@volatilesystems.org>